+++ omnist/__init__.py
@@ -75,7 +75,7 @@ from .errors import (
     WriteError,
 )
 
-__version__ = "0.1.8"
+__version__ = "0.1.9"
 
 __all__ = [
     # errors